Database.Sql.Type.Names

Documentation

type ConstrainSNames (c :: * -> Constraint) r a = (c a, c (TableRef r a), c (TableName r a), c (CreateTableName r a), c (DropTableName r a), c (SchemaName r a), c (CreateSchemaName r a), c (ColumnRef r a), c (NaturalColumns r a), c (UsingColumn r a), c (StarReferents r a), c (ITablishAliases r a), c (OTablishAliases r a), c (PositionExpr r a), c (ComposedQueryColumns r a))

type ConstrainSASNames (c :: (* -> *) -> Constraint) r = (c (TableRef r), c (TableName r), c (CreateTableName r), c (DropTableName r), c (SchemaName r), c (CreateSchemaName r), c (ColumnRef r), c (NaturalColumns r), c (UsingColumn r), c (StarReferents r), c (ITablishAliases r), c (OTablishAliases r), c (PositionExpr r), c (ComposedQueryColumns r))

class Resolution r

Associated Types

type TableRef r :: * -> *

TableRef refers to either a table in the catalog, or an alias

type TableName r :: * -> *

TableName refers to a table in the catalog

type CreateTableName r :: * -> *

CreateTableName refers to a table that might be in the catalog

Used for CREATE TABLE, special rules for resolution

type DropTableName r :: * -> *

DropTableName refers to a table that might be in the catalog

Used for DROP TABLE, special rules for resolution

type SchemaName r :: * -> *

SchemaName refers to a schema in the catalog

type CreateSchemaName r :: * -> *

CreateSchemaName refers to a table that might be in the catalog

Used for CREATE SCHEMA, special rules for resolution

type ColumnRef r :: * -> *

ColumnRef refers to either a column in the catalog, or an alias

type NaturalColumns r :: * -> *

NaturalColumns refers to columns compared in a natural join

type UsingColumn r :: * -> *

UsingColumn refers to columns that appear in USING (...)

type StarReferents r :: * -> *

type OTablishAliases r :: * -> *

OTablishAliases - Optional aliases decorating a Tablish

type ITablishAliases r :: * -> *

ITablishAliases - Implied aliases decorating a Tablish

Used where there's no space for adding aliases explicitly

type PositionExpr r :: * -> *

type ComposedQueryColumns r :: * -> *

type FQCN = FullyQualifiedColumnName

data FullyQualifiedColumnName

Constructors

FullyQualifiedColumnName

Fields

fqcnDatabaseName :: Text

fqcnSchemaName :: Text

fqcnTableName :: Text

fqcnColumnName :: Text

type FQTN = FullyQualifiedTableName

data FullyQualifiedTableName

Constructors

FullyQualifiedTableName

Fields

fqtnDatabaseName :: Text

fqtnSchemaName :: Text

fqtnTableName :: Text

qualifyColumnName :: FQTableName a -> UQColumnName b -> FQColumnName ()

fqcnToFQCN :: FQColumnName a -> FullyQualifiedColumnName

fqtnToFQTN :: FQTableName a -> FullyQualifiedTableName

data DatabaseName a

Constructors

DatabaseName a Text

data SchemaType

Constructors

NormalSchema

SessionSchema

data QSchemaName f a

Constructors

QSchemaName

Fields

schemaNameInfo :: a

schemaNameDatabase :: f (DatabaseName a)

schemaNameName :: Text

schemaNameType :: SchemaType

type UQSchemaName = QSchemaName No

type OQSchemaName = QSchemaName Maybe

type FQSchemaName = QSchemaName Identity

mkNormalSchema :: Alternative f => Text -> a -> QSchemaName f a

arbitraryUnquotedIdentifier :: Gen Text

arbitraryQuotedIdentifier :: Gen Text

arbitraryIdentifier :: Gen Text

data Identifier (ids :: [Symbol])

Identifiers picked first from (and shrunk to) symbols in type list. Used for testing.

Constructors

Identifier Text

class KnownSymbols (xs :: [Symbol]) where

Minimal complete definition

symbolVals

Methods

symbolVals :: proxy xs -> [String]

data QTableName f a

Constructors

QTableName

Fields

tableNameInfo :: a

tableNameSchema :: f (QSchemaName f a)

tableNameName :: Text

data No a

Constructors

None

type UQTableName = QTableName No

type OQTableName = QTableName Maybe

type FQTableName = QTableName Identity

newtype TableAliasId

Constructors

TableAliasId Integer

data TableAlias a

Constructors

TableAlias a Text TableAliasId

tableAliasName :: TableAlias a -> UQTableName a

data RNaturalColumns a

Constructors

RNaturalColumns [RUsingColumn a]

data RUsingColumn a

Constructors

RUsingColumn (RColumnRef a) (RColumnRef a)

data QFunctionName f a

Constructors

QFunctionName

Fields

functionNameInfo :: a

functionNameSchema :: f (QSchemaName f a)

functionNameName :: Text

type FunctionName = QFunctionName Maybe

data QColumnName f a

Constructors

QColumnName

Fields

columnNameInfo :: a

columnNameTable :: f (QTableName f a)

columnNameName :: Text

type UQColumnName = QColumnName No

type OQColumnName = QColumnName Maybe

type FQColumnName = QColumnName Identity

newtype ColumnAliasId

Constructors

ColumnAliasId Integer

data ColumnAlias a

Constructors

ColumnAlias a Text ColumnAliasId

columnAliasName :: ColumnAlias a -> UQColumnName a

columnAliasId :: ColumnAlias a -> ColumnAliasId

data RColumnRef a

Constructors

RColumnRef (FQColumnName a)

RColumnAlias (ColumnAlias a)

data StructFieldName a

Constructors

StructFieldName a Text

newtype FieldChain

Constructors

FieldChain (Map (StructFieldName ()) FieldChain)

data ParamName a

Constructors

ParamName a Text