Database.Sql.Type.Query

Documentation

data Query r a

Constructors

QuerySelect a (Select r a)

QueryExcept a (ComposedQueryColumns r a) (Query r a) (Query r a)

QueryUnion a Distinct (ComposedQueryColumns r a) (Query r a) (Query r a)

QueryIntersect a (ComposedQueryColumns r a) (Query r a) (Query r a)

QueryWith a [CTE r a] (Query r a)

QueryOrder a [Order r a] (Query r a)

QueryLimit a (Limit a) (Query r a)

QueryOffset a (Offset a) (Query r a)

newtype Distinct

Constructors

Distinct Bool

notDistinct :: Distinct

data CTE r a

Constructors

CTE

Fields

cteInfo :: a

cteAlias :: TableAlias a

cteColumns :: [ColumnAlias a]

cteQuery :: Query r a

data Select r a

Constructors

Select

Fields

selectInfo :: a

selectCols :: SelectColumns r a

selectFrom :: Maybe (SelectFrom r a)

selectWhere :: Maybe (SelectWhere r a)

selectTimeseries :: Maybe (SelectTimeseries r a)

selectGroup :: Maybe (SelectGroup r a)

selectHaving :: Maybe (SelectHaving r a)

selectNamedWindow :: Maybe (SelectNamedWindow r a)

selectDistinct :: Distinct

data SelectColumns r a

Constructors

SelectColumns

Fields

selectColumnsInfo :: a

selectColumnsList :: [Selection r a]

data SelectFrom r a

Constructors

SelectFrom a [Tablish r a]

data OptionalTablishAliases a

Constructors

TablishAliasesNone

TablishAliasesT (TableAlias a)

TablishAliasesTC (TableAlias a) [ColumnAlias a]

data RTablishAliases a

Constructors

RTablishAliases (Either TableAliasId (TableAlias a)) [ColumnAlias a]

data Tablish r a

Constructors

TablishTable a (OTablishAliases r a) (TableRef r a)

TablishSubQuery a (OTablishAliases r a) (Query r a)

TablishJoin a (ITablishAliases r a) (JoinType a) (JoinCondition r a) (Tablish r a) (Tablish r a)

TablishLateralView a (OTablishAliases r a) (LateralView r a) (Maybe (Tablish r a))

data JoinType a

Constructors

JoinInner a

JoinLeft a

JoinRight a

JoinFull a

JoinSemi a

data JoinCondition r a

Constructors

JoinNatural a (NaturalColumns r a)

JoinOn (Expr r a)

JoinUsing a [UsingColumn r a]

data LateralView r a

Constructors

LateralView

Fields

lateralViewInfo :: a

lateralViewOuter :: Maybe a

lateralViewExprs :: [Expr r a]

lateralViewWithOrdinality :: Bool

data SelectWhere r a

Constructors

SelectWhere a (Expr r a)

data SelectTimeseries r a

Constructors

SelectTimeseries

Fields

selectTimeseriesInfo :: a

selectTimeseriesSliceName :: ColumnAlias a

selectTimeseriesInterval :: Constant a

selectTimeseriesPartition :: Maybe (Partition r a)

selectTimeseriesOrder :: Expr r a

data PositionOrExpr r a

Constructors

PositionOrExprPosition a Int (PositionExpr r a)

PositionOrExprExpr (Expr r a)

data GroupingElement r a

Constructors

GroupingElementExpr a (PositionOrExpr r a)

GroupingElementSet a [Expr r a]

data SelectGroup r a

Constructors

SelectGroup

Fields

selectGroupInfo :: a

selectGroupGroupingElements :: [GroupingElement r a]

data SelectHaving r a

Constructors

SelectHaving a [Expr r a]

data SelectNamedWindow r a

Constructors

SelectNamedWindow a [NamedWindowExpr r a]

data Order r a

Constructors

Order a (PositionOrExpr r a) (OrderDirection (Maybe a)) (NullPosition (Maybe a))

data OrderDirection a

Constructors

OrderAsc a

OrderDesc a

data NullPosition a

Constructors

NullsFirst a

NullsLast a

NullsAuto a

data Offset a

Constructors

Offset a Text

data Limit a

Constructors

Limit a Text

data Selection r a

Constructors

SelectStar a (Maybe (TableRef r a)) (StarReferents r a)

SelectExpr a [ColumnAlias a] (Expr r a)

data Constant a

Constructors

StringConstant a ByteString
nb: Encoding *probably* matches server encoding, but there are ways to cram arbitrary byte sequences into strings on both Hive and Vertica.

NumericConstant a Text

NullConstant a

BooleanConstant a Bool

TypedConstant a Text (DataType a)

ParameterConstant a

data DataTypeParam a

Constructors

DataTypeParamConstant (Constant a)

DataTypeParamType (DataType a)

data DataType a

Constructors

PrimitiveDataType a Text [DataTypeParam a]

ArrayDataType a (DataType a)

MapDataType a (DataType a) (DataType a)

StructDataType a [(Text, DataType a)]

UnionDataType a [DataType a]

data Operator a

Constructors

Operator Text

data ArrayIndex a

Constructors

ArrayIndex a Text

data Expr r a

Constructors

BinOpExpr a (Operator a) (Expr r a) (Expr r a)

CaseExpr a [(Expr r a, Expr r a)] (Maybe (Expr r a))

UnOpExpr a (Operator a) (Expr r a)

LikeExpr a (Operator a) (Maybe (Escape r a)) (Pattern r a) (Expr r a)

ConstantExpr a (Constant a)

ColumnExpr a (ColumnRef r a)

InListExpr a [Expr r a] (Expr r a)

InSubqueryExpr a (Query r a) (Expr r a)

BetweenExpr a (Expr r a) (Expr r a) (Expr r a)

OverlapsExpr a (Expr r a, Expr r a) (Expr r a, Expr r a)

FunctionExpr a (FunctionName a) Distinct [Expr r a] [(ParamName a, Expr r a)] (Maybe (Filter r a)) (Maybe (OverSubExpr r a))

AtTimeZoneExpr a (Expr r a) (Expr r a)

SubqueryExpr a (Query r a)

ArrayExpr a [Expr r a]

ExistsExpr a (Query r a)

FieldAccessExpr a (Expr r a) (StructFieldName a)

ArrayAccessExpr a (Expr r a) (Expr r a)

TypeCastExpr a CastFailureAction (Expr r a) (DataType a)

VariableSubstitutionExpr a

data CastFailureAction

Constructors

CastFailureToNull

CastFailureError

newtype Escape r a

Constructors

Escape

Fields

escapeExpr :: Expr r a

newtype Pattern r a

Constructors

Pattern

Fields

patternExpr :: Expr r a

data Filter r a

Constructors

Filter

Fields

filterInfo :: a

filterExpr :: Expr r a

data Partition r a

Constructors

PartitionBy a [Expr r a]

PartitionBest a

PartitionNodes a

data FrameType a

Constructors

RowFrame a

RangeFrame a

data FrameBound a

Constructors

Unbounded a

CurrentRow a

Preceding a (Constant a)

Following a (Constant a)

data Frame a

Constructors

Frame

Fields

frameInfo :: a

frameType :: FrameType a

frameStart :: FrameBound a

frameEnd :: Maybe (FrameBound a)

data OverSubExpr r a

Constructors

OverWindowExpr a (WindowExpr r a)

OverWindowName a (WindowName a)

OverPartialWindowExpr a (PartialWindowExpr r a)

data WindowExpr r a

Constructors

WindowExpr

Fields

windowExprInfo :: a

windowExprPartition :: Maybe (Partition r a)

windowExprOrder :: [Order r a]

windowExprFrame :: Maybe (Frame a)

data PartialWindowExpr r a

Constructors

PartialWindowExpr

Fields

partWindowExprInfo :: a

partWindowExprInherit :: WindowName a

partWindowExprPartition :: Maybe (Partition r a)

partWindowExprOrder :: [Order r a]

partWindowExprFrame :: Maybe (Frame a)

data WindowName a

Constructors

WindowName a Text

data NamedWindowExpr r a

Constructors

NamedWindowExpr a (WindowName a) (WindowExpr r a)

NamedPartialWindowExpr a (WindowName a) (PartialWindowExpr r a)

arbitraryDate :: Gen Text

arbitraryTime :: Gen Text

arbitraryTimestamp :: Gen Text

arbitraryInterval :: Gen Text

arbitraryText :: Gen Text

arbitraryByteString :: Gen ByteString

shrinkByteString :: ByteString -> [ByteString]

scaleDown :: Int -> Gen a -> Gen a