Database.Sql.Util.Eval.Concrete

Documentation

data Concrete

data SqlValue

Constructors

SqlInt Integer

SqlStr ByteString

SqlBool Bool

SqlStruct (Map (StructFieldName ()) SqlValue)

SqlNull

truthy :: SqlValue -> Bool

truthy tells us whether our Haskell code should consider a SqlValue "true", mostly for the purpose of filtering. It should not be used to construct a SqlValue, as it does not handle NULL correctly for that purpose.

extendWithNulls :: [a] -> [SqlValue] -> [SqlValue]