Database.Sql.Util.Lineage.Table

Documentation

type TableLineage = Map FQTN (Set FQTN)

TableLineage is a set of descendants, each with an associated set of ancestors. Ancestors, for each descendant table, should contain a superset of all proximate tables that could have had an impact on the contents of the descendant following execution of the statement.

class HasTableLineage q where

Minimal complete definition

getTableLineage

Methods

getTableLineage :: q -> TableLineage

mkFQTN :: FQTableName a -> FullyQualifiedTableName

emptyLineage :: FullyQualifiedTableName -> TableLineage

squashTableLineage :: TableLineage -> TableLineage -> TableLineage

tableLineage :: Statement d ResolvedNames a -> TableLineage